Move vectorized version into package.
- Take the data that was in tensorflow and make a new complex
  package for it.

- Have test properly reference test data.
